ndlocr-api/requirements.txt
# Dependencies for ndlocr-lite API
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12

# ndlocr-lite dependencies
dill==0.3.8
lxml==5.4.0
networkx==3.3
onnxruntime==1.23.2
pillow==12.1.1
ordered-set==4.1.0
protobuf==6.31.1
pyparsing==3.1.2
PyYAML==6.0.1
tqdm==4.66.4
reportlab==4.2.5
pypdfium2==4.30.0
numpy==2.2.2
opencv-python-headless==4.11.0.86
